Moving tests to existing project
