+++ nanoFramework.CoreLibrary.sln
@@ -69,8 +69,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "nanoFramework.TestAdapter",
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "NFUnitTestGC", "Tests\NFUnitTestGC\NFUnitTestGC.nfproj", "{768BE3B6-41C6-4DFB-8A2D-443B2113F5AD}"
 EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "NFUnitTestString", "Tests\NFUnitTestString\NFUnitTestString.nfproj", "{60149C13-29FF-431A-A95A-91F31286EE80}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -231,12 +229,6 @@ Global
 		{768BE3B6-41C6-4DFB-8A2D-443B2113F5AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{768BE3B6-41C6-4DFB-8A2D-443B2113F5AD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{768BE3B6-41C6-4DFB-8A2D-443B2113F5AD}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Release|Any CPU.Build.0 = Release|Any CPU
-		{60149C13-29FF-431A-A95A-91F31286EE80}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -267,7 +259,6 @@ Global
 		{0BE498D1-CB3E-4D1E-BA4C-2C49AE30432D} = {0BAE286A-5434-4F56-A9F1-41B72056170E}
 		{6D740F0A-D435-4ACF-AD27-D702A599F229} = {0BAE286A-5434-4F56-A9F1-41B72056170E}
 		{768BE3B6-41C6-4DFB-8A2D-443B2113F5AD} = {0BAE286A-5434-4F56-A9F1-41B72056170E}
-		{60149C13-29FF-431A-A95A-91F31286EE80} = {0BAE286A-5434-4F56-A9F1-41B72056170E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8DE44407-9B41-4459-97D2-FCE54B1F4300}
